moved show graphs to separate script
