+++ src/frontend/Assets/Scripts/ShowGraphs.cs
@@ -1,0 +1,22 @@
+using CommunicationProtocol.Receivers;
+using CommunicationProtocol.Senders;
+using UnityEngine;
+using UnityEngine.UI;
+
+class GraphButton : MonoBehaviour
+{
+    [SerializeField] private Button graphButton;
+
+    private PythonRunner pythonRunner = new PythonRunner(PathConstants.PYTHON_ENVIRONMENT_PATH);
+
+    private void Start()
+    {
+        graphButton.onClick.AddListener(ShowGraphs);
+    }
+
+    public void ShowGraphs()
+    {
+        InputParameters inputParameters = new InputParameters(PathConstants.INPUT_PARAMETERS_PATH);
+        pythonRunner.Run(PathConstants.GRAPH_SCRIPT_PATH, inputParameters);
+    }
+}
